refactor: move constant to root and pretify imports
