+++ src/App.tsx
@@ -1,7 +1,8 @@
 import { BrowserRouter as Router, Routes } from 'react-router-dom';
-import { ErrorBoundary } from 'react-error-boundary';
 import { routes } from './routes';
+
 import { Footer, Header, FallBackUI } from './components';
+import { ErrorBoundary } from 'react-error-boundary';
 import { ToastContainer } from 'react-toastify';
 import { LanguageProvider } from './contexts/LanguageProvider';
