+++ src/components/Footer.tsx
@@ -1,6 +1,8 @@
 import { useContext } from 'react';
 import { translations } from '../contexts/translations';
+
 import { LanguageContext } from '../contexts/LanguageContext';
+
 import { PiGithubLogoFill } from 'react-icons/pi';
 import courseLogo from '../assets/rs_school_js.svg';
