+++ src/pages/NotFoundPage.tsx
@@ -1,8 +1,9 @@
 import { useContext } from 'react';
-import { Link } from 'react-router-dom';
 import { translations } from '../contexts/translations';
 import { LanguageContext } from '../contexts/LanguageContext';
-import { MEMES_URL } from '../utils/constants';
+import { MEMES_URL } from '../constants';
+
+import { Link } from 'react-router-dom';
 
 import '@styles/NotFoundPage.css';
